Changes to use np.float32 instead of np.float

+++ src/cuda_kernel_runner/arg_handler.py
@@ -17,7 +17,7 @@ type_conv_dict = {
     "longlong": np.longlong,
     "ulonglong": np.ulonglong,
     "half": np.half,
-    "float": np.float,
+    "float": np.float32,
     "double": np.float64,
     "longdouble": np.longdouble,
     "csingle": np.csingle,
@@ -28,23 +28,23 @@ type_conv_dict = {
 custom_type_dict = {
     "float2": {
         "names": ['x', 'y'],
-        "types": [np.float] * 2,
+        "types": [np.float32] * 2,
         "length": 2,
-        "repr_type": np.float
+        "repr_type": np.float32
     },
 
     "float3": {
         "names": ['x', 'y', 'z'],
-        "types": [np.float] * 3,
+        "types": [np.float32] * 3,
         "length": 3,
-        "repr_type": np.float
+        "repr_type": np.float32
     },
 
     "float4": {
         "names": ['x', 'y', 'z', 'w'],
-        "types": [np.float] * 4,
+        "types": [np.float32] * 4,
         "length": 4,
-        "repr_type": np.float
+        "repr_type": np.float32
     },
 
     "double2": {
